rename AssetPickerModal to ImageAssetPickerModal
