+++ src/plugins/dam/components/image-asset-picker-modal/index.jsx
